Moved PyramusConsts to framework
