+++ framework/src/main/java/fi/otavanopisto/pyramus/PyramusConsts.java
@@ -1,4 +1,4 @@
-package fi.otavanopisto.pyramus.rest.util;
+package fi.otavanopisto.pyramus;
 
 public class PyramusConsts {
